rename folder to make it work for Swiper 5.x
