+++ types/swiper/tsconfig.json
@@ -19,7 +19,7 @@
     },
     "files": [
         "index.d.ts",
-        "dist/js/swiper.esm.d.ts",
+        "js/swiper.esm.d.ts",
         "swiper-tests.ts"
     ]
 }
